Move deser/result to scylla-cql-core
